% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_perform_checks.R
\name{mod_perform_checks_ui}
\alias{mod_perform_checks_ui}
\alias{mod_perform_checks_server}
\title{mod_perform_checks_ui and mod_perform_checks_server}
\usage{
mod_perform_checks_ui(id)

mod_perform_checks_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
